+++ src/inmanta/db/versions/v2.py
@@ -15,7 +15,6 @@ ALTER TABLE public.report ALTER COLUMN completed DROP NOT NULL;
 
 CREATE INDEX compile_env_requested_index ON compile (environment, requested ASC);
 CREATE INDEX compile_env_remote_id_index ON compile (environment, remote_id);
-
 """
     async with connection.transaction():
         await connection.execute(schema)